Move notification test to notifications test suite
