+++ tests/Feature/Checkins/AssetCheckinTest.php
@@ -9,9 +9,7 @@ use App\Models\LicenseSeat;
 use App\Models\Location;
 use App\Models\Statuslabel;
 use App\Models\User;
-use App\Notifications\CheckinAssetNotification;
 use Illuminate\Support\Facades\Event;
-use Illuminate\Support\Facades\Notification;
 use Tests\Support\InteractsWithSettings;
 use Tests\TestCase;
 
@@ -168,53 +166,5 @@ class AssetCheckinTest extends TestCase
         Event::assertDispatched(function (CheckoutableCheckedIn $event) {
             return $event->action_date === '2023-01-02 12:45:56' && $event->note === 'hello';
         }, 1);
-    }
-
-    public function testCheckInEmailSentToUserIfSettingEnabled()
-    {
-        Notification::fake();
-
-        $user = User::factory()->create();
-        $asset = Asset::factory()->assignedToUser($user)->create();
-
-        $asset->model->category->update(['checkin_email' => true]);
-
-        event(new CheckoutableCheckedIn(
-            $asset,
-            $user,
-            User::factory()->checkinAssets()->create(),
-            ''
-        ));
-
-        Notification::assertSentTo(
-            [$user],
-            function (CheckinAssetNotification $notification, $channels) {
-                return in_array('mail', $channels);
-            },
-        );
-    }
-
-    public function testCheckInEmailNotSentToUserIfSettingDisabled()
-    {
-        Notification::fake();
-
-        $user = User::factory()->create();
-        $asset = Asset::factory()->assignedToUser($user)->create();
-
-        $asset->model->category->update(['checkin_email' => false]);
-
-        event(new CheckoutableCheckedIn(
-            $asset,
-            $user,
-            User::factory()->checkinAssets()->create(),
-            ''
-        ));
-
-        Notification::assertNotSentTo(
-            [$user],
-            function (CheckinAssetNotification $notification, $channels) {
-                return in_array('mail', $channels);
-            }
-        );
     }
 }
