+++ tests/Feature/Notifications/AssetWebhookTest.php
@@ -108,6 +108,54 @@ class AssetWebhookTest extends TestCase
         Notification::assertNotSentTo(new AnonymousNotifiable, CheckinAssetNotification::class);
     }
 
+    public function testCheckInEmailSentToUserIfSettingEnabled()
+    {
+        Notification::fake();
+
+        $user = User::factory()->create();
+        $asset = Asset::factory()->assignedToUser($user)->create();
+
+        $asset->model->category->update(['checkin_email' => true]);
+
+        event(new CheckoutableCheckedIn(
+            $asset,
+            $user,
+            User::factory()->checkinAssets()->create(),
+            ''
+        ));
+
+        Notification::assertSentTo(
+            [$user],
+            function (CheckinAssetNotification $notification, $channels) {
+                return in_array('mail', $channels);
+            },
+        );
+    }
+
+    public function testCheckInEmailNotSentToUserIfSettingDisabled()
+    {
+        Notification::fake();
+
+        $user = User::factory()->create();
+        $asset = Asset::factory()->assignedToUser($user)->create();
+
+        $asset->model->category->update(['checkin_email' => false]);
+
+        event(new CheckoutableCheckedIn(
+            $asset,
+            $user,
+            User::factory()->checkinAssets()->create(),
+            ''
+        ));
+
+        Notification::assertNotSentTo(
+            [$user],
+            function (CheckinAssetNotification $notification, $channels) {
+                return in_array('mail', $channels);
+            }
+        );
+    }
+
     private function createAsset()
     {
         return Asset::factory()->laptopMbp()->create();
